capstone-ecommerce-microservices/product-service/product_prj/requirements.txt
Django
djangorestframework
snowflake-connector-python
django-snowflake  
python-dotenv
django-cors-headers